Convert NetworkController deps to TS
We want to convert NetworkController to TypeScript in order to be able
to compare differences in the controller between in this repo and the
core repo. To do this, however, we need to convert the dependencies of
the controller to TypeScript.

To this end, this commit converts the following files:

- shared/modules/fetch-with-timeout.js
- shared/modules/network.utils.js
- shared/constants/metametrics.js

Note that NetworkController depends on some NPM modules that have no
type definitions. These will be addressed in a future commit.